AND-7905 Fixed UI.
